+++ shop/src/main/java/com/falcon/shop/domain/shop/Orders.java
@@ -38,11 +38,11 @@ public class Orders extends Base {
   private String guestEmail;                 // 비회원 이메일
   private String guestFirstName;             // 비회원 성
   private String guestLastName;              // 비회원 이름
-  private Long totalPrice;                   // 총 가격
+  private Double totalPrice;                   // 총 가격
   private Long totalQuantity;                // 총 수량
   private Long totalItemCount;               // 총 항목수
-  private Long shipPrice;                    // 배송비
-  private String paymentMethod;              // 결제방식 (CASH, COIN, CARD, TRANSFER)
+  private Double shipPrice;                    // 배송비
+  private String paymentMethod;              // 결제방식 (TRANSFER, COIN)
   private String status;                     // 상태 ('결제대기','결제완료','배송중','배송완료','주문취소','환불완료')
  
   @TableField(exist = false)
